Update filter to be class based
Update the filter to be classed based. This has benifits and has compacted the code as well. It also makes the module more plesant to use for the end user. Docstrings for public functions were also updated to be more short and to the point but still contain all of the information needed to use the functions. These changes also include updates to comply with PEP8 standards.